requirements.txt
streamlit==1.49.0
tensorflow-cpu==2.20.0
numpy>=1.26,<3
opencv-python
Pillow

